% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_to_poisson.R
\name{AdjustToPoisson}
\alias{AdjustToPoisson}
\title{Adjust to poisson}
\usage{
AdjustToPoisson(randomIntersections, pAdjust = "fdr")
}
\arguments{
\item{randomIntersections}{This is a data frame with each row containing
a number of overlap for each category in the columns.}

\item{pAdjust}{Correction method. Can be abbreviated.}

\item{showCategories=TRUE}{Show the categories when they are computed.
Make debugging easier.}
}
\value{
A data frame containing all the chi2 test informations.
}
\description{
Adjust random tests to the poisson distribution
through goodness of fit test whith chi2.
}
\examples{
catalog <- BedToGranges(system.file("extdata", "ReMap_nrPeaks_public_chr22.bed",
                          package = "roken"))
randoms <- RandomIntersections(catalog, 500, 1000, 1000)
chi2results <- AdjustToPoisson(randoms)

}
\author{
Zacharie Menetrier
}
