% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_import.R
\name{bedImport}
\alias{bedImport}
\title{Import bed files}
\usage{
bedImport(path)
}
\arguments{
\item{path}{The path of the file to import.}
}
\value{
A data frame with the corresponding chromosomic regions 
of the bed file.
}
\description{
Imports bed files automatically in a data frame.
}
\examples{
chromFile <- system.file("extdata", "hg19.genome", package = "roken")
chromSizes <- bedImport(chromFile)

}
\author{
Zacharie Menetrier
}
