% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_granges.R
\name{BedToGranges}
\alias{BedToGranges}
\title{Bed to GRanges}
\usage{
BedToGranges(path)
}
\arguments{
\item{path}{Location of your file.}
}
\value{
A genomic ranges object containing the informations from the given
bed file.
}
\description{
This function loads a bed file and stores it as a genomic 
ranges object.
The tab-delimited file must be ordered as 'chrom', 'chromStart', 
'chromEnd', 'name', 'score', 'strand'.
The minimal bed file must have the 'chrom', 'chromStart', 
'chromEnd' columns.
Any columns after the strand column are ignored.
}
\examples{
queryFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22_SOX2.bed",
                         package = "roken")
granges <- BedToGranges(queryFile)

}
\author{
Martin Mestdagh
}
