% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_enrichment_bed.R
\name{category.enrichment.bed}
\alias{category.enrichment.bed}
\title{Compute categorical enrichment}
\usage{
category.enrichment.bed(regions, catalog, genome, fraction = 0.1,
  shuffles = 10)
}
\arguments{
\item{regions}{The bed file containing the genomic regions to analyze.}

\item{catalog}{The bed file containing the database used for annotation.}

\item{genome}{A file containing all the chromosome lengths for the species in consideration.}

\item{fraction}{The fraction that intersections shall exceed to be considered.}

\item{shuffles}{The number of shuffled genomic regions to be created for theorical distribution (higher means more accurate).}
}
\value{
A data frame containing the enrichment informations.
}
\description{
Gets the value of genomic enrichment for each feature of a bed file.
}
