% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment.R
\name{computeEnrichment}
\alias{computeEnrichment}
\title{Computes enrichment}
\usage{
computeEnrichment(query, catalog, chromSizes, fractionQuery,
  fractionCatalog, shufflesNumber, categories, universe, byChrom, included,
  nCores)
}
\arguments{
\item{query}{The object containing the genomic regions to analyze.}

\item{catalog}{The object containing the database used for annotation.}

\item{chromSizes}{An object containing all the chromosome lengths for the 
species in consideration.}

\item{fractionQuery}{The fraction of coverage (query on catalog) a hit must
exceed to be accounted.}

\item{fractionCatalog}{The fraction of coverage (catalog on query) a hit must
exceed to be accounted.}

\item{categories}{The categories contained in the catalog.
This option is leaved for faster calculation when this function is runned
multiple times.}

\item{shuffles}{The number of shuffled genomic regions to be created for 
theorical distribution (higher means more accurate).}

\item{universe=NULL}{A set of genomic regions that prevent shuffles
for occuring outside of it.}

\item{byChrom=FALSE}{Will the shuffles stay in the chromosome they originate (TRUE)
or can they be placed everywhere on the genome (FALSE)}

\item{included=1}{Represents the fraction of each regions that can
be outside of the universe.}
}
\value{
A list containing the number of overlaps for the query and the 
mean number of overlaps for the shuffles.
}
\description{
Gets the number of overlaps for the query and the shuffles for
 each category of the catalog.
}
\author{
Zacharie Menetrier
}
