% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_overlaps.R
\name{count.overlaps}
\alias{count.overlaps}
\title{Count overlaps}
\usage{
count.overlaps(bedfile1, bedfile2, fraction = 0.1)
}
\arguments{
\item{bedfile1}{The first bed file.}

\item{bedfile2}{The second bed file.}

\item{fraction}{The fraction that intersections shall exceed to be considered.}
}
\value{
The number of overlaps for the two sets of chromosomic regions given a frequency.
}
\description{
Gets the number of overlaps for two chromosomic regions sets at a certain frequency.
The chromosomic regions are read from bed files.
}
