% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mapped_peaks.R
\name{CountMappedPeaks}
\alias{CountMappedPeaks}
\title{Count mapped peaks}
\usage{
CountMappedPeaks(query, catalog)
}
\arguments{
\item{query}{A genomic ranges object containing the peaks to analyze.}

\item{catalog}{A genomic ranges object containing the peaks that will be 
compared to the query.}
}
\value{
An integer that represents the number of peaks mapped.
}
\description{
This function counts the number of peaks that are mapped with 
two genomic ranges object.
It counts the number of peaks in the query that are mapped in the catalog.
}
\examples{
queryFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22_SOX2.bed",
                          package = "roken")
catalogFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22.bed",
                            package = "roken")
query <- BedToGranges(queryFile)
catalog <- BedToGranges(catalogFile)
countMappedPeaks <- CountMappedPeaks(query, catalog)

}
\author{
Zacharie Menetrier
}
