% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_remap_catalog.R
\name{downloadRemapCatalog}
\alias{downloadRemapCatalog}
\title{Download Remap catalogue}
\usage{
downloadRemapCatalog(targetDir, fileName = "", 
version = "2018", assembly = "hg38", force = FALSE, store = TRUE)
}
\arguments{
\item{targetDir}{The name of the directory to download the catalogue in.}

\item{fileName=""}{The name of the file after downloading.
If left empty, default url names will be applied.}

\item{version="2018"}{The year version of the catalog 2018 or 2015.}

\item{assembly="hg38"}{The  genomic version of assembly hg38 or hg19 for human.}

\item{force=FALSE}{If FALSE (default), then no file is overwrited and the
user is given confirmation message.}

\item{store=TRUE}{If TRUE (default) then a file is downloaded and written
on the disk else it is only loaded as an R object.}
}
\value{
A data frame containing the ReMap genomic regions if store = FALSE
else the path to the catalog file.
}
\description{
Download or import in a data frame the Remap catalogue for 
 transcriptions factors.
}
\examples{
remapFile <- downloadRemapCatalog(targetDir = "~/ReMapEnrich_demo/data/ReMap")
remap <- bedToGranges(remapFile)

}
\author{
Zacharie Menetrier
}
