% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ucsc_chrom_sizes.R
\name{downloadUcscChromSizes}
\alias{downloadUcscChromSizes}
\title{Download ucsc chromosomes sizes}
\usage{
downloadUcscChromSizes(id, all = FALSE)
}
\arguments{
\item{id}{The ucsc id of the assembly.}

\item{all=FALSE}{If TRUE then all the chromosomes sizes will be returned in
the data frame, only the ones with no "_" in their names will be returned.}
}
\value{
A vector that contains the chromosome lengths.
}
\description{
Download and import the chromosome sizes from any species in
the ucsc data base.
}
\examples{
rn5ChromSizes <- donwloadUcscsChromSizes("rn5")

}
\author{
Zacharie Menetrier
}
