% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment}
\usage{
enrichment(query, catalog, chromSizes = LoadChromSizes("hg19"),
                    fractionQuery = 0.1, fractionCatalog = 0.1,
                    shuffles = 6, lower = FALSE, pAdjust = "BY",
                     byChrom = FALSE, included = 1, nCores = 1)
}
\arguments{
\item{query}{The genomic ranges object containing the genomic regions to analyze.}

\item{catalog}{The genomic ranges object containing the database used for annotation.}

\item{chromSizes=LoadChromSizes("hg19")}{A vector containing all the chromosome 
lengths for the species in consideration.}

\item{fractionQuery=0.1}{The fraction of coverage (query on catalog) a hit must 
exceed to be accounted.}

\item{fractionCatalog=0.1}{The fraction of coverage (catalog on query) a hit
must exceed to be accounted.}

\item{shuffles=6}{The number of shuffled genomic regions to be created for 
theorical distribution (higher means more accurate).}

\item{tail="lower"}{If "lower" then, probabilities are P[X > x], 
if "higher", P[X <= x], if "both" then higher or lower is selected
depending on the number of overlaps vs the theorical mean.}

\item{pAdjust="BY"}{The method that will be used for correcting the p-values.
BH, BY and bonferroni are available.}

\item{universe=NULL}{A set of genomic regions that prevent shuffles
for occuring outside of it.}

\item{included=1}{Represents the fraction of each regions that can
be outside of the universe.}

\item{byChrom=FALSE}{Will the shuffles stay in the chromosome they originate (TRUE)
or can they be placed everywhere on the genome (FALSE)}

\item{nCores=1}{The number of cores to be used for parallel computations.
1 by default will not be using parallel calculations.}
}
\value{
A data frame containing the enrichment informations.
}
\description{
Gets the value of genomic enrichment for each category of a
genomic ranges object against a catalogue.
}
\examples{
queryFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22_SOX2.bed", 
                         package = "roken")
catalogFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22.bed",
                           package = "roken")
query <- bedToGranges(queryFile)
catalog <- bedToGranges(catalogFile)
enrichment <- enrichment(query, catalog)

}
\author{
Zacharie Menetrier
}
