% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_dot_plot.R
\name{enrichmentDotPlot}
\alias{enrichmentDotPlot}
\title{Enrichment dot plot}
\usage{
enrichmentDotPlot(enrich, 
top = 20,
main = paste("Significance, top", top, "categories"),
sigType = "q",
col = c("#6699ff", "#ff5050"),
minCircleSize = 0.5,
inches = 1 / 4,
xlab    = "Mapped peaks ratio",
stainCol = rgb(colorFunction(((sig - minSigLim) / (maxSigLim - minSigLim))),
 maxColorValue = 255),
...)
}
\arguments{
\item{enrich}{The enrichment data frame from which the plot will be created.}

\item{op=20}{The number of category for the plot.}

\item{main=paste("Significance, }{top", top, "categories"). Allows to choose
the title of the plot.}

\item{aRisk=0.05}{The alpha risk.}

\item{col=c("#6699ff", "#ff5050")}{Palette of coloration for the plot 
Personnal coloration such as c("#FEE0D2","#FC9272") or a RColorBrewer 
such as brewer.pal(5,"Reds").}

\item{minCircleSize=0.5}{Change the minimum size of circle for the plot.}

\item{inches=1/4}{Modify the area of points.}

\item{xlab="Mapped}{peaks ratio" Allows to change the title of x-axis.}

\item{stainCol=rgb(colorFunction(((sig}{- minSigLim) /
(maxSigLim - minSigLim))), maxColorValue = 255)
Changes the color of perimeter of each point.}
}
\description{
Creates a dotplot from the enrichment(category=f(q.significance))
}
\examples{
data("enrichment_example", package = "roken")
enrichmentDotPlot(enrichment_example)

}
\author{
Martin Mestdagh
}
