% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_volcano_plot.R
\name{enrichmentVolcanoPlot}
\alias{enrichmentVolcanoPlot}
\title{Enrichment volcano plot}
\usage{
enrichmentVolcanoPlot(enrich,
main = "Volcano plot",
aRisk = 0.05,
sigDisplayQuantile = 0.95,
col = c("#BFCBD4", "#33ACFF"),
sigType = "q",
ylim = c(0,yMax),
xlab = "Effect size",
ylab = sigTypeTitle,
pch = pch,
cex = 0.8,
las = 1,
...)
}
\arguments{
\item{enrich}{The file enrichment from which the plot will be create.}

\item{main=c("Volcanoplot}{of category") change the plot title.}

\item{aRisk=0.05}{The alpha risk, by default 0.05.}

\item{sigDisplayQuantile=0.95}{quantile used to define the maximal value for
the Y axis, based on a quantile.}

\item{col=c("#6699ff", "#ff5050")}{Palette of coloration for the histogram 
with personnal color or RColorBrewer palette.}

\item{sigType="q"}{Allows to choose between Q-significance or P-significance
or E-significance.}

\item{ylim=c(0, yMax)}{Create the ylim with the quantile selected.}

\item{xlab="Effect}{size" Allows to change label of x-axis.}

\item{ylab="Significance"}{Allows to change label of y-axis.}

\item{pch=pch}{Allows to choose shape of points outside quantile.}

\item{cex=0.8}{Allows to choose the diamater of the points.}

\item{las=1}{Allows to change the angle of label y-axis.}
}
\description{
Creates a volcano plot from the enrichment.
}
\examples{
data("enrichment_example", package = "ReMapEnrich")
enrichmentVolcanoPlot(na.omit(enrichment_example))

}
\author{
Martin Mestdagh
}
