% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_enrichment.R
\name{extractEnrichment}
\alias{extractEnrichment}
\title{Extract enrichment}
\usage{
extractEnrichment(categories, tail, categoriesOverlaps, theoricalMeans,
  categoriesCount, pAdjust = "BH")
}
\arguments{
\item{tail}{If "lower" then, probabilities are P[X > x], 
if "higher", P[X <= x], if "both" then higher or lower is selected
depending on the number of overlaps vs the theorical mean.}

\item{categoriesOverlaps}{The number of overlaps for each category.}

\item{theoricalMeans}{The mean number of overlaps for each category.}

\item{categoriesCount}{A vector representing the number of time a category 
is found in the catalog.}

\item{pAdjust}{method  used to correct p-values for multiple testing. Supported: "BH" (Benjamini-Hochberg), "BY", "bonferroni".}

\item{categories=unique(catalog@elementMetadata$id)}{The categories contained in the catalog.
This option is leaved for faster calculation when this function 
is runned multiple times.}
}
\value{
A data frame containing the enrichment informations.
}
\description{
Extracts the information for genomic enrichment.
}
\author{
Zacharie Menetrier
}
