% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_regions.R
\name{GenRegions}
\alias{GenRegions}
\title{Generate regions}
\usage{
GenRegions(n, size, chromSizes = LoadChromSizes("hg19")), 
universe = NULL, included = 1
}
\arguments{
\item{n}{The number of regions.}

\item{size}{The width of the regions.}

\item{chromSizes=ImportChromSizes("hg19")}{The chromosome sizes for the 
species in consideration.}

\item{universe=NULL}{A set of genomic regions that prevent random regions
for occuring outside of it.}

\item{included=1}{Represents the fraction of each regions that can
be outside of the universe.}
}
\value{
A random generated genomic ranges object.
}
\description{
Generates random genomic ranges object.
}
\examples{
randomRegions <- GenRegions(1000,1000)

}
\author{
Zacharie Menetrier
}
