% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_to_bed.R
\name{grangesToBed}
\alias{grangesToBed}
\title{Granges to bed}
\usage{
grangesToBed <- function(granges)
}
\arguments{
\item{granges}{The genomic ranges object to convert.}
}
\value{
A data frame containing the informations from the given 
genomic ranges object.
}
\description{
This function creates a bed data frame from a 
genomic ranges object.
The data frame will be ordered as 'chrom', 'chromStart', 'chromEnd',
'name', 'score', 'strand'.
}
\examples{
catalogFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22.bed", 
                            package = "roken")
catalog <- bedToGranges(catalogFile)
catalog <- grangesToBed(catalogFile)
View(catalog)

}
\author{
Zacharie Menetrier
}
