% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bed.R
\name{import.bed}
\alias{import.bed}
\title{Import bed files}
\usage{
import.bed(path)
}
\arguments{
\item{path}{The path of the file to import.}
}
\value{
A data frame with the corresponding chromosomic regions of the bed file.
}
\description{
Imports bed files automatically in a nice data frame.
}
