% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chrom_sizes.R
\name{ImportChromSizes}
\alias{ImportChromSizes}
\title{Import chromosomes sizes}
\usage{
ImportChromSizes <- function(filePath)
}
\arguments{
\item{filePath}{The path of the file containing the chromsomes sizes.}
}
\value{
A table that contains the chromosome sizes.
}
\description{
Import chromosome sizes file into a table.
}
\examples{
hg19ChromFile <- LoadChromFile("hg19")
hg19ChromSizes <- ImportChromSizes(hg19ChromFile)

}
\author{
Zacharie Menetrier
}
