% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_bed.R
\name{intersect.bed}
\alias{intersect.bed}
\title{Genomic regions intersections}
\usage{
intersect.bed(bedfile1, bedfile2, fraction = 0.1)
}
\arguments{
\item{bedfile1}{The first bed file.}

\item{bedfile2}{The second bed file.}

\item{fraction}{The fraction that intersections shall exceed to be considered.}
}
\value{
A data frame (empty if no intersections have been found) containing the intersections and their lengths in base pairs.
}
\description{
Gets the intersections of two chromosomic regions at a certain fraction.
The chromosomic regions are read from bed files.
}
