% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{intersect}
\alias{intersect}
\title{Intersect}
\usage{
intersect(query, catalog, fractionQuery = 0.1, fractionCatalog = 0.1,
                   categories = unique(catalog@elementMetadata$id))
}
\arguments{
\item{query}{A genomic ranges object to be analyzed.}

\item{catalog}{A genomic ranges object to be compared to the query.}

\item{fractionQuery=0.1}{The fraction of coverage (query on catalog) a hit
must exceed to be accounted.}

\item{fractionCatalog=0.1}{The fraction of coverage (catalog on query) a hit
must exceed to be accounted.}

\item{categories=unique(catalog@elementMetadata$id)}{The categories 
contained in the catalog.
This option is leaved for faster calculation when this function is runned 
multiple times.}
}
\value{
A vector containing the numbers of hits for each category of the catalog.
}
\description{
Computes the intersections between two genomic ranges object.
}
\examples{
queryFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22_SOX2.bed",
                          package = "roken")
catalogFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22.bed",
                            package = "roken")
query <- bedToGranges(queryFile)
catalog <- bedToGranges(catalogFile)
intersects <- intersect(query, catalog)

}
\author{
Zacharie Menetrier
}
