% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_file.R
\name{output.file}
\alias{output.file}
\title{Create an output file with the format chosen by user or a capture like 'txt' if none is selected}
\usage{
output.file(dat, name.file = "output.file", format = "default")
}
\description{
@param dat dataframe to output
 @param name.file Name of the output file chosen
 @param format The format chosen
     
 @export
}
