% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_intersections.R
\name{randomIntersections}
\alias{randomIntersections}
\title{Random intersections}
\usage{
randomIntersections(catalog, iterations, regionNb,
regionSize, chromSizes = LoadChromSizes("hg38")), nCores = "auto"
}
\arguments{
\item{catalog}{A GRanges object containing the regions to be overlapped with.}

\item{iterations}{The number of time the intersections will be computed.}

\item{regionNb=1000}{The number of regions to be created for each random query.}

\item{regionSize=1000}{The size of each regions for each random query.}

\item{universe=NULL}{A set of genomic regions that prevent random regions
for occuring outside of it.}

\item{included=1}{Represents the fraction of each regions that can
be outside of the universe.}

\item{chromSizes=LoadChromSizes("hg38")}{A vector containing all the chromosome 
lengths for the species in consideration.}

\item{byChrom=FALSE}{Will the shuffles/random stay in the chromosome they 
originate (TRUE) or can they be placed everywhere on the genome (FALSE)}

\item{shuffle=NULL}{A set of genomic regions to shuffle for random
intersections instead of complete randomly generated regions.}

\item{nCores="auto"}{The number of cores to be used for parallel computations.
By default it is the number of available cores minus one.}
}
\value{
A data frame with each row containing
a number of overlap for each category in the columns.
}
\description{
Creates random query regions
and returns the number of overlaps with a catalogue for each iteration.
}
\examples{
catalog <- bedToGranges(system.file("extdata", "ReMap_nrPeaks_public_chr22.bed",
                          package = "ReMapEnrich"))
randoms <- randomIntersections(catalog, 500, 1000, 1000)

}
\author{
Zacharie Menetrier
}
