% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roken.R
\docType{package}
\name{Roken}
\alias{roken}
\alias{roken-package}
\alias{Roken-package}
\title{roken}
\description{
Bioinformatic tools for computing enrichment
}
\author{
\strong{Maintainer}: Zacharie Menetrier \email{zacharie.menetrier@gmail.com}

Authors:
\itemize{
  \item Martin Mestdagh \email{martin.mestdagh@gmail.com}
}

}
