% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_bed.R
\name{shuffle.bed.temp.file}
\alias{shuffle.bed.temp.file}
\title{Genomic regions shuffle}
\usage{
shuffle.bed.temp.file(regions, genome)
}
\arguments{
\item{regions}{The bed file containing the genomic regions to shuffle.}

\item{genome}{A file containing all the chromosome lengths for the species in consideration.}
}
\value{
The path to a temporary file containing the new shuffled chromosic regions.
}
\description{
Shuffle genomic regions among a defined genome.
}
