% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(regions, chromSizes = LoadChromSizes("hg38"), universe = NULL,
included = 1, byChrom = FALSE)
}
\arguments{
\item{regions}{The genomic ranges object containing the genomic 
regions to shuffle.}

\item{chromSizes=LoadChromSizes("hg38")}{A vector containing all 
the chromosome lengths for the species in consideration.}

\item{universe=NULL}{A set of genomic regions that prevent shuffles
for occuring outside of it.}

\item{included=1}{Represents the fraction of each regions that can
be outside of the universe.}

\item{byChrom=FALSE}{Will the shuffles stay in the chromosome they originate (TRUE)
or can they be placed everywhere on the genome (FALSE)}
}
\value{
A genomic ranges containing the new shuffled chromosic regions.
}
\description{
Shuffle genomic regions among a defined genome.
}
\examples{
regionsFile <- system.file("extdata", "ReMap_nrPeaks_public_chr22.bed",
                            package = "roken")
regions <- bedToGranges(regionsFile)
shuffledRegions <- shuffle(regions)

}
\author{
Zacharie Menetrier
}
