% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_outline.R
\name{textOutline}
\alias{textOutline}
\title{Text outline}
\usage{
textOutline <- function(x, y, labels, col = 'white',
                               outline = 'black', ... )
}
\arguments{
\item{x, y}{numeric vectors of coordinates where the text
labels should be written.}

\item{labels}{a character vector or expression specifying the text 
to be written.}

\item{col}{The color to be used for the front text.}

\item{outline}{The color to be used for the outline.}
}
\description{
Draw text with an outline.
}
\author{
Zacharie Menetrier
}
