% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_enrichment.R
\name{exportEnrichment}
\alias{exportEnrichment}
\title{Export enrichment}
\usage{
exportEnrichment(enrichment, fileName = "output", format = "tsv")
}
\arguments{
\item{enrichment}{Enrichment data frame to output.}

\item{fileName="output"}{Name of the output file.}

\item{format="tsv"}{The format of the output file.}
}
\description{
Create an output file for an enrichment data frame.
}
\examples{
data("enrichment_example", "ReMapEnrich")
exportEnrichment(enrichment_example)
     
}
\author{
Martin Mestdagh
}
