% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chrom_sizes.R
\name{importChromSizes}
\alias{importChromSizes}
\title{Import chromosomes sizes}
\usage{
importChromSizes <- function(filePath)
}
\arguments{
\item{filePath}{The path of the file containing the chromsomes sizes.}
}
\value{
A table that contains the chromosome sizes.
}
\description{
Import chromosome sizes file into a table.
}
\examples{
hg19ChromFile <- loadChromFile("hg19")
hg19ChromSizes <- importChromSizes(hg19ChromFile)

}
\author{
Zacharie Menetrier
}
