% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chrom_sizes.R
\name{LoadChromSizes}
\alias{LoadChromSizes}
\title{Load chromosomes sizes}
\usage{
LoadChromSizes <- function(genome)
}
\arguments{
\item{genome}{The name of the species to import the chromosomes from.}
}
\value{
A table that contains the chromosome lengths.
}
\description{
Load preloaded chromosome sizes file into a table.
}
\examples{
hg19ChromSizes <- LoadChromSizes("hg19")

}
\author{
Zacharie Menetrier
}
