% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loladb_to_catalog.R
\name{LolaDBToCatalog}
\alias{LolaDBToCatalog}
\title{Lola DB to catalogue}
\usage{
LolaDBToCatalog(regionDB)
}
\arguments{
\item{regionDB}{The region DB to convert.}
}
\value{
A GenomicRanges object.
}
\description{
Create a GenomicRanges object ready to be used as a catlogue
from Lola databases objects.
}
\author{
Zacharie Menetrier
}
